# docs/openapi/meetings.yaml
openapi: 3.0.3
info:
  title: UIP Meetings & Collaboration Platform API
  description: |
    Native, in-platform video conferencing and collaboration for UIP
    (Universities Innovation Platform) — meeting scheduling, lobby/guest
    access, real-time signaling, WebRTC (mesh, client-side), live chat,
    reactions, screen sharing, host controls, file sharing, notes, polls,
    attendance, bulk invitations + calendar, client-side recording, and
    admin analytics/monitoring.

    Every endpoint below returns the project-wide response envelope:

    ```json
    { "success": true|false, "message": "...", "data": ..., "errors": null, "meta": {} }
    ```

    `success=false` responses always use this envelope too, with `message`
    holding the human-readable error and `errors` holding field-level
    validation messages (Laravel Validator format) when applicable.
  version: "1.0.0"
  contact:
    name: UIP Platform Team
servers:
  - url: /api/v1
    description: Same-origin API (Laravel backend, mounted under /api/v1)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Standard UIP JWT bearer token (UipJwtService), same auth used across
        the whole platform. Endpoints under `uip.auth.optional` middleware
        accept either this header OR a `guest_token` (see below) — never
        both required.
  parameters:
    MeetingUuid:
      name: uuid
      in: path
      required: true
      description: The meeting's public UUID (never the internal numeric id).
      schema:
        type: string
        format: uuid
      example: 8f14e45f-ceea-467e-add6-01ab2cd39e12
...
  schemas:
    ApiEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: {}
        errors:
          type: object
          nullable: true
        meta:
          type: object
    ApiError:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
      example:
        success: false
        message: "You are not authorized to join this meeting."
        data: null
        errors: null
        meta: {}
...
    Meeting:
      type: object
      properties:
        id: { type: integer, example: 4210 }
        uuid: { type: string, format: uuid, example: 8f14e45f-ceea-467e-add6-01ab2cd39e12 }
        host_user_id: { type: integer, example: 118 }
        title: { type: string, example: "Graduation Project Meeting" }
        description: { type: string, nullable: true, example: "Weekly project discussion" }
        type: { type: string, enum: [instant, scheduled], example: scheduled }
        status: { type: string, enum: [scheduled, lobby, live, ended, cancelled], example: scheduled }
        scheduled_start_at: { type: string, format: date-time, nullable: true, example: "2026-09-05T18:00:00Z" }
        duration_minutes: { type: integer, example: 60 }
        waiting_room_enabled: { type: boolean, example: true }
        allow_guests: { type: boolean, example: false }
        max_participants: { type: integer, example: 8 }
        has_password: { type: boolean, example: false }
        locked: { type: boolean, example: false }
        settings:
          type: object
          example:
            mute_on_entry: false
            camera_on_entry: true
            allow_chat: true
            allow_file_sharing: true
            recording_enabled: true
        meeting_code:
          type: string
          nullable: true
          description: Only present when the response explicitly includes join details (create/show).
          example: "482-119-733"
        join_url:
          type: string
          nullable: true
          example: "https://uip.example.edu/meetings/join/jt_9k3mQ2xVh7pR4sT8"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
...
  /meetings/{uuid}/start:
    parameters: [{ $ref: '#/components/parameters/MeetingUuid' }]
    post:
      tags: [Meetings]
      summary: Start a scheduled/instant meeting (host only)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Meeting started successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/Meeting' } }
        '403':
          description: Only the host can start this meeting.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ApiError' } }
  /meetings/{uuid}/end:
    parameters: [{ $ref: '#/components/parameters/MeetingUuid' }]
    post:
      tags: [Meetings]
      summary: End a live meeting (host only)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Meeting ended successfully.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope' } }
        '403':
          description: Only the host can end this meeting.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ApiError' } }
...